requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
drf-yasg==1.21.7
filelock==3.15.4
inflection==0.5.1
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pytz==2024.1
PyYAML==6.0.1
setuptools==72.1.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
virtualenv==20.26.3
